[Fix] Fix reveal in value method.
