+++ contract/AElf.Contracts.Consensus.AEDPoS/AEDPoSContract_HelpMethods_GetInformationToUpdateConsensus.cs
@@ -44,9 +44,7 @@ namespace AElf.Contracts.Consensus.AEDPoS
                 outValue, signature);
 
             ShareInValueOfCurrentRound(updatedRound, previousRound, inValue, publicKey);
-
-            RevealSharedInValues(updatedRound, previousRound, publicKey);
-
+            
             // To publish Out Value.
             return new AElfConsensusHeaderInformation
             {
@@ -77,9 +75,18 @@ namespace AElf.Contracts.Consensus.AEDPoS
         private AElfConsensusHeaderInformation GetInformationToUpdateConsensusForNextRound(Round currentRound,
             string publicKey, AElfConsensusTriggerInformation triggerInformation)
         {
-            Assert(
-                GenerateNextRoundInformation(currentRound, Context.CurrentBlockTime, out var nextRound),
-                "Failed to generate next round information.");
+            if (!GenerateNextRoundInformation(currentRound, Context.CurrentBlockTime, out var nextRound))
+            {
+                Assert(false, "Failed to generate next round information.");
+            }
+
+            RevealSharedInValues(currentRound, publicKey);
+
+            if (!TryToUpdateRoundInformation(currentRound))
+            {
+                Assert(false, "Failed to update current round information after revealing previous in values.");
+            }
+
             nextRound.RealTimeMinersInformation[publicKey].ProducedBlocks =
                 nextRound.RealTimeMinersInformation[publicKey].ProducedBlocks.Add(1);
             Context.LogDebug(() => $"Mined blocks: {nextRound.GetMinedBlocks()}");
